Add readAllBytes2 utility method
