+++ acornui-core/jvm/src/main/kotlin/com/acornui/jvm/loader/JvmBinaryLoader.kt
@@ -19,10 +19,10 @@ package com.acornui.jvm.loader
 import com.acornui.core.asset.AssetType
 import com.acornui.io.NativeReadByteBuffer
 import com.acornui.jvm.io.JvmByteBuffer
+import com.acornui.jvm.io.readAllBytes2
 import java.io.InputStream
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
-import java.util.*
 
 open class JvmBinaryLoader(
 		path: String,
@@ -40,43 +40,5 @@ open class JvmBinaryLoader(
 		val buffer = ByteBuffer.wrap(byteArray)
 		buffer.order(ByteOrder.LITTLE_ENDIAN)
 		return JvmByteBuffer(buffer)
-	}
-
-	/**
-	 * Jvm 9's readAllBytes method, but for Jvm 6+
-	 */
-	private fun InputStream.readAllBytes2(): ByteArray {
-		var buf = ByteArray(DEFAULT_BUFFER_SIZE)
-		var capacity = buf.size
-		var nread = 0
-		var n: Int
-		while (true) {
-			// read to EOF which may read more or less than initial buffer size
-			while (true) {
-				n = read(buf, nread, capacity - nread)
-				if (n <= 0) break
-				nread += n
-			}
-
-			// if the last call to read returned -1, then we're done
-			if (n < 0)
-				break
-
-			// need to allocate a larger buffer
-			if (capacity <= MAX_BUFFER_SIZE - capacity) {
-				capacity = capacity shl 1
-			} else {
-				if (capacity == MAX_BUFFER_SIZE)
-					throw OutOfMemoryError("Required array size too large")
-				capacity = MAX_BUFFER_SIZE
-			}
-			buf = Arrays.copyOf(buf, capacity)
-		}
-		return if (capacity == nread) buf else Arrays.copyOf(buf, nread)
-	}
-
-	companion object {
-
-		private const val MAX_BUFFER_SIZE = Integer.MAX_VALUE - 8
 	}
 }
